app/requirements.txt
python-binance==1.0.29
python-dotenv==1.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
sqlparse>=0.4.4,<0.5
sqlalchemy>=2.0.0,<2.1
alembic>=1.13,<1.14
plotly==5.21.0
dash==2.17.0
pydantic==2.11.7
apscheduler==3.11.0